test(ssr): move SSR test to separate 'ssr.test.tsx' file
